+++ packages/toasts/src/index.ts
@@ -1,0 +1,7 @@
+import '@tokene/styles/src/index.scss'
+import 'vue-toastification/src/scss/index.scss'
+
+export * from './components'
+export * from './composables'
+export * from './types'
+export { default as ToastPlugin } from 'vue-toastification'